+++ src/Plugins/Nop.Plugin.Payments.PayPalStandard/plugin.json
@@ -2,7 +2,7 @@
   "Group": "Payment methods",
   "FriendlyName": "PayPal Standard",
   "SystemName": "Payments.PayPalStandard",
-  "Version": "1.63",
+  "Version": "1.64",
   "SupportedVersions": [ "4.40" ],
   "Author": "nopCommerce team",
   "DisplayOrder": 1,
